+++ src/main/kotlin/crimera/patches/twitter/misc/recommendedusers/fingerprints/HideRecommendedUsersFingerprint.kt
@@ -1,4 +1,4 @@
-package crimera.patches.twitter.recommendedusers.fingerprints
+package crimera.patches.twitter.misc.recommendedusers.fingerprints
 
 import app.revanced.patcher.fingerprint.MethodFingerprint
 import com.android.tools.smali.dexlib2.Opcode